Speed improvements
So my get neighbour algorithim was naive and poorly performing, because it evaluated ALL of the cells in the entire grid.

Shocking.

So, I've adapted the neighbour counting algorithm from http://rosettacode.org/wiki/Conway's_Game_of_Life#C.23. It's much better.

I've also removed the Cell class and changed the grid to be an array of bools.

Overall it's a nice speedup.